Several updates to the download feature
Rename the variable download_folder to download_output as it can contain a prefix to the downloaded files names

When setting the download_output, we check if the target folder is writable. If it is not, the output is not changed.
On the same line, the download mode cannot be enabled if the current download_output is not writable.
We test for 'dirname "$download_output"X' so that whether there is a prefix or not (/tmp/ytdl- vs /tmp/), we test the right folder and never the parent folder

We also check, before starting any download, if there is more than min_avail_space free space (default is 1go, has to be changed in the conf file directly)
If not, the video is not downloaded and autoplay is disabled